+++ lib/phoenix_feathers/components/button.ex
@@ -1,9 +1,17 @@
 defmodule PhoenixFeathers.Button do
+  @doc """
+  Example
+
+  ```
+  <%= live_component @socket, PhoenixFeathers.Button, display_text: "Click me" %>
+  ```
+  """
   use PhoenixFeathers.LiveComponent
 
   def render(%{click_event: click_event} = assigns) do
     ~L"""
       <button
+        class="phx_feathers_button"
         phx-click="<%= click_event %>"
       >
         <%= @display_text %>
@@ -13,7 +21,7 @@ defmodule PhoenixFeathers.Button do
 
   def render(assigns) do
     ~L"""
-      <button>
+      <button class="phx_feathers_button">
         <%= @display_text %>
       </button>
     """