+++ lib/phoenix_feathers/components/card.ex
@@ -1,10 +1,21 @@
 defmodule PhoenixFeathers.Card do
+  @doc """
+  Example
+
+  ```
+  <%= live_component @socket, PhoenixFeathers.Card do %>
+    <div>My card content</div>
+  <% end %>
+  ```
+  """
+
   use PhoenixFeathers.LiveComponent
   use Phoenix.HTML
 
   def render(%{inner_block: inner_block, click_event: click_event} = assigns) do
     ~L"""
       <div
+        class="phx_feathers_card"
         phx-click="<%= click_event %>"
       >
         <%= render_block(inner_block) %>
@@ -14,7 +25,7 @@ defmodule PhoenixFeathers.Card do
 
   def render(%{inner_block: inner_block} = assigns) do
     ~L"""
-      <div>
+      <div class="phx_feathers_card">
         <%= render_block(inner_block) %>
       </div>
     """