+++ test/phoenix_feathers_test.exs
@@ -1,8 +1,4 @@
 defmodule PhoenixFeathersTest do
   use ExUnit.Case
   doctest PhoenixFeathers
-
-  test "greets the world" do
-    assert PhoenixFeathers.hello() == :world
-  end
 end