+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.0-alpha]
+### Added
+* Basic side bar component
+* Basic top nav component
+* Mix tasks
+* SCSS file that can be copied into a project using `mix phx.feathers --styles`
+
 ## [0.1.0-alpha]
 ### Added
 * Basic button component